slack_bot/requirements.txt
pandas
requests
pandas
sqlalchemy
python-dotenv
psycopg2-binary